// txcnotes/source/txc-front-web-toc/src/app/pages/merchant/models/merchant.model.ts
import { MerchantAddress } from './merchant-address.model'; 

export interface Merchant {
    merchantId: number;
    programId: number;
    programCode: number;
    identityCode: string;
    merchantName: string;
    externalCode: string;
    invoiceRegisterNumber: string;
    securityKey: string;
    status: number;
    tX1MerchantUID: string;
    isAutoCreateReimbursement: boolean;
    autoCreateReimbursementIntervalType: number;
    reimbursementTaxType: number;
    reimbursementType: number;
    reimbursementReceivers: string;
    sameKeyWithShop: boolean;
    mamEmail: string;
    merchantAutoType: number;
    notificationMerchantCode: string;
    merchantEmail: string;
    merchantAcquireId: number;
    merchantAcquirerId: number;
    needConsumerScan: boolean;
    memo: string;
    issuerType: number;
    workKeyId: number;
    workKey: null;
    workKeyExpireTime: string;
    workKeyCreatedTime: string;
    isLegacyMerchant: boolean;
    mainContact: string;
    mainPhone: string;
    autoCreateReimbursementDay: number;
    description: string;
    categoryId: number;
    address: MerchantAddress;
    merchantContactEmailList: string[] | any[];
    edenredContactEmailList: string[] | any[];
    merchantAddress?: string;
}